+++ dev/joss/paper.bib
@@ -28,7 +28,7 @@
   number        = {8},
   pages         = {741--756},
   publisher     = {Taylor \& Francis},
-  doi           = {10.1163/156939388X00044},
+  url           = {https://www.tandfonline.com/doi/abs/10.1163/156939388X00044},
 }
 
 @book{hansen1988,
@@ -57,7 +57,7 @@
   journaltitle  = {Electromechanics of Particles},
   pages         = {227–235},
   publisher     = {Cambridge University Press},
-  doi           = {https://doi.org/10.1017/CBO9780511574498}
+  doi           = {10.1017/CBO9780511574498}
 }
 
 @article{hofmannExcitationAwareSelfAdaptiveFrequency2023,
@@ -105,7 +105,7 @@
   pages     = {65--98},
   year      = {2017},
   publisher = {Society of Industrial and Applied Mathematics},
-  doi       = {https://doi.org/10.1137/141000671}
+  doi       = {10.1137/141000671}
 }
 
 @book{harringtonFieldComputationMoment1993,
