### 2023-08-01 #5717 @HoBeZwe
Hello, and thanks for handling my submission.

I fixed the DOIs. However, the first invalid one seems to be indeed provided incorrectly by the publisher. Hence, I removed it.
