### 2023-08-23 #5717 @iliarasskazov
Dear Pierre,With pleasure.Best,IliaOn Aug 23, 2023, at 6:53 AM, Pierre de Buyl ***@***.***> wrote:﻿
@iliarasskazov would you be willing to review this article for the Journal of Open Source Software?
