### 2023-09-08 #5717 @pdebuyl
Hi @berquist , it is often the case to have some tension between domain knowledge. At JOSS, the extra requirement on programming makes it a bit more frequent. I will assign you as a reviewer. If at some point in the review you feel that you need to reconsider let me know.
