### 2023-09-08 #5820 @editorialbot
[REVIEW]: SphericalScattering: A Julia Package for Electromagnetic Scattering from Spherical Objects
**Submitting author:** @HoBeZwe (<a href="http://orcid.org/0000-0003-1435-6203">Bernd Hofmann</a>)
**Repository:** https://github.com/HoBeZwe/SphericalScattering.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.6.0
