+++ ext/plotting.jl
@@ -1,0 +1,6 @@
+
+function SphericalScattering.plotFF()
+
+    plot(scatter(x=1:5, y=rand(5), mode="markers"))
+end
+
