+++ ext/SphericalScatteringExt.jl
@@ -1,0 +1,18 @@
+module SphericalScatteringExt
+
+# -------- used packages
+using SphericalScattering
+using PlotlyJS
+
+
+# -------- exportet parts
+# types
+
+# functions
+#export SphericalScattering.plotFF
+
+
+# -------- included files
+include("plotting.jl")
+
+end
