### 2023-09-24 4c9847a
implement FF plot + update documentation
