### 2023-09-25 d4ccec3
implement plotffcut
