### 2023-09-25 5feebe6
add near-field example for plotting
