+++ docs/src/application.md
@@ -94,7 +94,7 @@ print("Far-field error: $diff_FF %\n")
 
 
 ---
-## Visualization of Scattering 
+## Visualization of Fields 
 
 This package provides several means to directly visualize quantities of the scattering setup.
 
@@ -103,7 +103,7 @@ This package provides several means to directly visualize quantities of the scat
     (It is a [weak dependency](https://pkgdocs.julialang.org/v1/creating-packages/#Conditional-loading-of-code-in-packages-(Extensions)).)
 
 
-### Plot the Far-Field Patterns
+### Plotting Far-Field Patterns
 
 As an example consider a Hertzian dipole that excites a PEC sphere:
 
@@ -154,7 +154,7 @@ savefig(t, "plotPatternHDtot.html"); nothing # hide
 
 
 
-### Plot Field cuts
+### Plotting Far-Field Cuts
 
 Sphercial cuts can be conveniently obtained by:
 
@@ -186,4 +186,80 @@ savefig(t, "plotcut.html"); nothing # hide
 
 ```@raw html
 <object data="plotcut.html" type="text/html"  style="width:100%;height:50vh;"> </object>
+```
+
+
+### Plotting Near-Field Cuts
+
+The near-field of an electric ring current can, e.g., be visualized in the xz-plane as:
+
+```@example heatmap
+using SphericalScattering
+using LinearAlgebra, StaticArrays
+
+f = 1e8             # frequency
+c = 2.99792458e8    # speed of light
+λ = c / f           # wavelength
+
+ex = electricRingCurrent(frequency=1e8, center=SVector(0.,0,0), radius=3*λ)
+
+# --- define points in the xz plane
+res = λ/15
+
+points_cart = [SVector(x, 0.0, z) for z in -5λ:res:5λ, x in -5λ:res:5λ]
+points_sph = SphericalScattering.cart2sph.(points_cart)
+
+# --- evaluate the fields
+E = field(ex, ElectricField(points_cart))
+
+nothing # hide
+```
+
+To plot the fields a heatmap can be employed:
+
+```@example heatmap
+using PlotlyJS
+
+# --- convert data to spherical coordinates
+Esph = SphericalScattering.convertCartesian2Spherical.(E, points_sph)
+Eφ = [Esph[i][3] for (i, j) in enumerate(Esph)] # extract φ-component
+
+# --- truncate large values (for plot)
+data = abs.(Eφ)
+data[data .> 250] .= 250
+
+# --- plot heatmap
+layout = Layout(
+    yaxis=attr(title_text="y/λ"),
+    xaxis=attr(title_text="x/λ")
+)
+
+plot(heatmap(x=-5:1/15:5,y=-5:1/15:5,z=data, colorscale="Jet"), layout)
+t = plot(heatmap(x=-5:1/15:5,y=-5:1/15:5,z=data, colorscale="Jet"), layout) # hide
+savefig(t, "plotNF.html"); nothing # hide
+```
+
+```@raw html
+<object data="plotNF.html" type="text/html"  style="width:60%;height:50vh;"> </object>
+```
+
+Or instead of the magnitude a snapshot can be plotted
+
+```@example heatmap
+data = real.(Eφ)
+data[data .> 250] .= 250
+
+# --- plot heatmap
+layout = Layout(
+    yaxis=attr(title_text="y/λ"),
+    xaxis=attr(title_text="x/λ")
+)
+
+plot(heatmap(x=-5:1/15:5,y=-5:1/15:5,z=data, colorscale="Jet"), layout)
+t = plot(heatmap(x=-5:1/15:5,y=-5:1/15:5,z=data, colorscale="Jet"), layout) # hide
[…]
