### 2023-09-25 6ef6c6f
update tests for new functionality + update manual
