### 2023-09-25 0b2b46a
update package dependencies for docs
