+++ docs/Project.toml
@@ -2,4 +2,6 @@
 BEAST = "bb4162c7-ba94-5a20-af32-d8ec4428bdd1"
 CompScienceMeshes = "3e66a162-7b8c-5da0-b8f8-124ecd2c3ae1"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a"
 SphericalScattering = "1a9ea918-b599-4f1f-bd9a-d681e8bb5b3e"
+StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
