### 2023-09-27 bebea6b
update joss references
