+++ dev/joss/paper.bib
@@ -87,6 +87,42 @@
   doi         = {10.23919/EuCAP57121.2023.10133154},
 }
 
+@inproceedings{hofmannLowFrequencyStableDiscretization2021,
+  title       = {Low-Frequency Stable Discretization of the Electric Field Integral Equation Based on Poincaré's Lemma},
+  booktitle   = {Proc. IEEE Antennas Propag. Soc. Int. Symp. URSI Nat. Radio Sci. Meeting},
+  author      = {Hofmann, Bernd and Eibert, Thomas F. and Andriulli, Francesco P. and Adrian, Simon B.},
+  date        = {2021-12},
+  publisher   = {IEEE},
+  location    = {Singapore},
+  doi         = {10.1109/APS/URSI47566.2021.9703799}
+}
+
+@inproceedings{hofmannLowFrequencyStabilizedElectricField2022,
+  title       = {Low-Frequency-Stabilized Electric Field Integral Equation on Topologically Non-Trivial Geometries for Arbitrary Excitations},
+  booktitle   = {Proc. IEEE Antennas Propag. Soc. Int. Symp. URSI Nat. Radio Sci. Meeting},
+  author      = {Hofmann, Bernd and Eibert, Thomas F. and Andriulli, Francesco P. and Adrian, Simon B.},
+  date        = {2022-07},
+  pages       = {1938--1939},
+  doi         = {10.1109/AP-S/USNC-URSI47032.2022.9886833}
+}
+
+@inproceedings{hofmannEfficientCombinationScalarPotential2022a,
+  title       = {Efficient Combination of Scalar-Potential Representations of Solenoidal Functions and Quasi-Helmholtz Projectors},
+  booktitle   = {16th European Conference on Antennas and Propagation (EuCAP)},
+  author      = {Hofmann, Bernd and Eibert, Thomas F. and Andriulli, Francesco P. and Adrian, Simon B.},
+  date        = {2022-03},
+  doi         = {10.23919/EuCAP53622.2022.9769650}
+}
+
+@inproceedings{hofmannSelfAdaptiveFrequencyNormalization2023,
+  title       = {Towards a Self-Adaptive Frequency Normalization Scheme for the Low-Frequency Stabilized Magnetic Field Integral Equation},
+  booktitle   = {Proc. IEEE Antennas Propag. Soc. Int. Symp. URSI Nat. Radio Sci. Meeting},
+  author      = {Hofmann, Bernd and Eibert, Thomas F. and Andriulli, Francesco P. and Adrian, Simon B.},
+  date        = {2023-07},
+  pages       = {1213--1214},
+  doi         = {10.1109/USNC-URSI52151.2023.10238214}
+}
+
 @article{adrianElectromagneticIntegralEquations2021,
   title         = {Electromagnetic Integral Equations: Insights in Conditioning and Preconditioning},
   author        = {Adrian, Simon B. and Dély, Alexandre and Consoli, Davide and Merlini, Adrien and Andriulli, Francesco P.},
@@ -178,4 +214,92 @@
   publisher   = {GitHub},
   journal     = {GitHub repository},
   url         = {https://github.com/JuliaWaveScattering/EffectiveWaves.jl}
+}
+
+@misc{matscat2023,
+  author      = {Jan Schäfer},
+  title       = {MatScat},
+  year        = {2023},
+  publisher   = {MATLAB},
+  journal     = {MATLAB Central File Exchange},
+  url         = {https://www.mathworks.com/matlabcentral/fileexchange/36831-matscat}
+}
+
+@misc{matmie2023,
+  author      = {Nicholas Walter},
+  title       = {Mie Electric Field Simulation for Spheres},
+  year        = {2023},
[…]
