### 2023-09-27 #5820 @HoBeZwe
Thank you already for everybody's effort to review so far.

We have added a section [Application Examles](https://hobezwe.github.io/SphericalScattering.jl/dev/application/#Application-Examples) to the documentation to further improve the usability and also to make clearer what kind of tests are performed.

We have also updated the paper and discuss now a wider range of related software including implementations in C++, Python, and MATLAB.

Concerning the Zenodo reference: To our understanding, it will be added to the paper only after the review (in case it is positive of course). But please correct me if I am wrong here.
