### 2023-09-28 #5820 @berquist
@HoBeZwe Could you say something here about some of the commented-out test code such as https://github.com/HoBeZwe/SphericalScattering.jl/blob/bebea6b63d8b9eef1f7c108e51d4cb6d06c57559/test/sphericalModes.jl#L39?
