### 2023-09-28 #5820 @berquist
For the summary, it isn't clear how the section starting "While there are some applications where canonical scattering problems (...)" is relevant at this particular point in the text, but it seems like it should be connected to the previous point about implementations of semi-analytical solutions.  Is that what this package implements, therefore enabling scattering calculations as part of more involved techniques?  I think the first paragraph of the Statement of Need says this, so some movement of text between these two parts might clarify the summary.
