### 2023-09-29 e76c5dd
stabilize normalization + fix bugs
