+++ src/sphericalModes/incident.jl
@@ -52,6 +52,7 @@ function field(excitation::SphericalModeTE, point, quantity::ElectricField; para
 
     # --- common factor
     pf = prefac(m, n)
+    ZF = sqrt(ε / μ)
 
     # --- Hankel functions
     if excitation.c == 1        # inward
@@ -63,21 +64,12 @@ function field(excitation::SphericalModeTE, point, quantity::ElectricField; para
     end
 
     # --- factors Eϑ and Eϕ have in common
-    mabs = abs(m)
-    aux =
-        excitation.amplitude *
-        k *
-        sqrt(sqrt(μ / ε)) *
-        H *
-        pf *
-        exp(-im * m * ϕ) *
-        sqrt((2 * n + 1) / 2 * factorial(n - mabs) / factorial(n + mabs))
+    aux = excitation.amplitude * k * sqrt(ZF) * H * pf * exp(im * m * ϕ)
 
     # --- put things together
     Eϑ = aux * im * associatedLegendre(n, m, ϑ)
-    Eϕ = aux * derivatieAssociatedLegendre(n, m, ϑ)
+    Eϕ = -aux * derivatieAssociatedLegendre(n, m, ϑ)
 
-    #return SVector(Er, Eϑ, Eϕ)
     return convertSpherical2Cartesian(SVector{3,Complex{T}}(Er, Eϑ, Eϕ), point_sph) # convert to Cartesian representation
 end
 
@@ -113,6 +105,7 @@ function field(excitation::SphericalModeTM, point, quantity::ElectricField; para
 
     # --- common factor
     pf = prefac(m, n)
+    ZF = sqrt(ε / μ)
 
     # --- Hankel functions
     if excitation.c == 1        # inward
@@ -130,19 +123,13 @@ function field(excitation::SphericalModeTM, point, quantity::ElectricField; para
 
     # --- factors Er, Eϑ, and Eϕ have in common
     mabs = abs(m)
-    aux =
-        excitation.amplitude *
-        sqrt(sqrt(μ / ε)) *
-        pf *
-        exp(-im * m * ϕ) *
-        sqrt((2 * n + 1) / 2 * factorial(n - mabs) / factorial(n + mabs)) / r
+    aux = excitation.amplitude * sqrt(ZF) * pf * exp(im * m * ϕ) / r
 
     # --- put things together
-    Er = aux * n * (n + 1) * H * (-1)^mabs * Plm(cos(ϑ), n, mabs)
+    Er = aux * n * (n + 1) * H * Plm(cos(ϑ), n, mabs; norm=Val(:normalized))
     Eϑ = aux * dH * derivatieAssociatedLegendre(n, m, ϑ)
-    Eϕ = -aux * dH * im * associatedLegendre(n, m, ϑ)
+    Eϕ = aux * dH * im * associatedLegendre(n, m, ϑ)
 
-    #return SVector(Er, Eϑ, Eϕ)
     return convertSpherical2Cartesian(SVector{3,Complex{T}}(Er, Eϑ, Eϕ), point_sph) # convert to Cartesian representation
 end
 
@@ -176,23 +163,15 @@ function field(excitation::SphericalModeTE, point, quantity::FarField; parameter
 
     # --- common factor
     pf = prefac(m, n)
+    ZF = sqrt(ε / μ)
 
     # --- factors Eϑ and Eϕ have in common
-    mabs = abs(m)
-    aux =
-        -excitation.amplitude *
-        (im)^(n + 1) *
-        T(1.0) *
[…]
