+++ test/sphericalModes.jl
@@ -9,87 +9,78 @@ T = assemble(𝑇, RT, RT)
 
 @testset "TE modes" begin
 
-    #f = 1e8
-    #κ = 2π * f / c   # Wavenumber
-
-    ex = SphericalModeTE(; frequency=f, m=0, n=1, c=2)
-
     @testset "Incident fields" begin
 
+        ex = SphericalModeTE(; frequency=f, m=0, n=1, c=2)
+
         # define an observation point
-        point_cart = [SVector(4.0, 2.0, 3.2)]
+        point_cart = [SVector(0.0, 1.0, 3.2)]
 
-        @test_nowarn EF = field(ex, ElectricField(point_cart))
-        @test_nowarn HF = field(ex, MagneticField(point_cart))
-        @test_nowarn FF = field(ex, FarField(point_cart))
+        EF = field(ex, ElectricField(point_cart))
+        HF = field(ex, MagneticField(point_cart))
+        FF = field(ex, FarField(point_cart))
 
-        point_cart = [SVector(0.0, 0.0, 3.2), SVector(0.0, 0.0, -3.2)]
+        # rather poor test: values are checked against values determined by this package
+        #                   only to see whether future changes result in the same values
+        @test FF[1][1] ≈ 0.005309361037222248 + 0.0im
+        @test FF[1][2] ≈ -3.2510459998759485e-19 + 0.0im
+        @test FF[1][3] ≈ 0.0 + 0.0im
 
-        ex2 = SphericalModeTE(; frequency=f, m=1, n=1, c=1)
+        @test EF[1][1] ≈ 0.0010133416507273605 - 0.0012376919299125465im
+        @test EF[1][2] ≈ -6.204928045029786e-20 + 7.578677301489551e-20im
+        @test EF[1][3] ≈ 0.0 + 0.0im
 
-        @test_nowarn EF = field(ex2, ElectricField(point_cart))
-        @test_nowarn HF = field(ex2, MagneticField(point_cart))
-        @test_nowarn FF = field(ex2, FarField(point_cart))
+        @test HF[1][1] ≈ 9.889059303803533e-23 - 2.333915413016954e-22im
+        @test HF[1][2] ≈ 1.6150059446835905e-6 - 3.811573124009171e-6im
+        @test HF[1][3] ≈ -3.6397996407963353e-6 - 1.3757077858915473e-6im
+
 
         ex3 = SphericalModeTE(; frequency=f, m=0, n=1, c=3)
 
         @test_throws ErrorException("Type can only be 1 or 2.") EF = field(ex3, ElectricField(point_cart))
         @test_throws ErrorException("Type can only be 1 or 2.") HF = field(ex3, MagneticField(point_cart))
-
-        # @test FF[1][1] ≈ 295.0240872654143 + 112.00825545163434im
-        # @test FF[1][2] ≈ 295.0240872654143 + 112.00825545163434im
-        # @test FF[1][3] ≈ -368.78010908176793 - 140.01031931454293im
-
-        # @test EF[1][1] ≈ 33.75576840234728 - 66.33989918223259im
-        # @test EF[1][2] ≈ 33.75576840234728 - 66.33989918223259im
-        # @test EF[1][3] ≈ -118.11657839808761 + 218.13026261521526im
-
-        # @test H[1][1] ≈ -0.23911545354270725 + 0.4457313513337625im
-        # @test H[1][2] ≈ 0.23911545354270725 - 0.4457313513337625im
-        # @test H[1][3] ≈ 0.0 + 0.0im
     end
 
     @testset "Scattered fields" begin
 
-        # ----- BEAST solution
-        𝐸 = ex
[…]
