+++ docs/src/sphModes.md
@@ -15,14 +15,14 @@
 <br/>
 ```
 
-!!! danger
-    This excitation needs substantial revision. So far only the low order modes are stable. Also, the superposition of multiple modes should be included.
+!!! info
+    So far only single modes (not their superposition) are implemented.
 
 
 ---
 ## Definition
 
-The spherical modes are defined following the conventions of [[5]](@ref refs).
+The spherical modes are defined following the conventions of [[5]](@ref refs), however adapted to the time convention ``\mathrm{e}^{\,\mathrm{j}\omega t}``.
 
 #### TE Modes
 
@@ -30,8 +30,8 @@ The functions
 ```math
 \begin{aligned}
 \bm{f}_{1mn}^{(c)}(r,\vartheta, \varphi) = \cfrac{1}{2\pi} \cfrac{1}{n(n+1)}{\left(-\frac{m}{|m|}\right)}^m 
-													&\left(-\mathrm{h}_n^{(c)}(kr)\frac{\mathrm{j}m\bar{\mathrm{P}}_n^{|m|}(\cos\vartheta)}{\sin\vartheta}\mathrm{e}^{-\mathrm{j}m\varphi} \bm{e}_\vartheta   	\right.	 \\	
-													&~~\left. -\mathrm{h}_n^{(c)}(kr)\cfrac{\mathrm{d} \bar{\mathrm{P}}_n^{|m|}(\cos\vartheta) }{\mathrm{d}\vartheta} \mathrm{e}^{-\mathrm{j}m\varphi} \bm{e}_\varphi  	~\right)
+													&\left(\mathrm{h}_n^{(c)}(kr)\frac{\mathrm{j}m\bar{\mathrm{P}}_n^{|m|}(\cos\vartheta)}{\sin\vartheta}\mathrm{e}^{\mathrm{j}m\varphi} \bm{e}_\vartheta   	\right.	 \\	
+													&~~\left. -\mathrm{h}_n^{(c)}(kr)\cfrac{\mathrm{d} \bar{\mathrm{P}}_n^{|m|}(\cos\vartheta) }{\mathrm{d}\vartheta} \mathrm{e}^{\mathrm{j}m\varphi} \bm{e}_\varphi  	~\right)
 \end{aligned}
 ```
 form the ``\mathrm{TE}_{mn}`` modes, where ``\mathrm{P}_n^m`` denote the associated Legendre polynomials and ``\mathrm{h}_n^{(c)}(x)`` denote the spherical Hankel functions.
@@ -42,13 +42,17 @@ form the ``\mathrm{TE}_{mn}`` modes, where ``\mathrm{P}_n^m`` denote the associa
 The functions
 ```math
 \begin{aligned}
-\bm{f}_{2mn}^{(c)}(r,\vartheta, \varphi) = \cfrac{1}{2\pi} \cfrac{1}{n(n+1)}{\left(-\frac{m}{|m|}\right)}^m \Bigg(&\cfrac{n(n+1)}{kr}\mathrm{h}_n^{(c)}   \bar{\mathrm{P}}_n^{|m|}(\cos\vartheta)	\mathrm{e}^{-\mathrm{j}m\varphi} \bm{e}_r \Bigg. \\	
[…]
