### 2023-09-29 #5820 @HoBeZwe
> @HoBeZwe Could you say something here about some of the commented-out test code such as https://github.com/HoBeZwe/SphericalScattering.jl/blob/bebea6b63d8b9eef1f7c108e51d4cb6d06c57559/test/sphericalModes.jl#L39?

Since the higher-order modes were not stabilized (as pointed out in the documentation) we were a bit lazy in the tests for this excitation. So a good reason to be skeptical here.

To properly fix this, we decided to directly remove the limitation and then upgrade the tests. The latter are now more thorough (especially, the [TE scattering case](https://github.com/HoBeZwe/SphericalScattering.jl/blob/47393e9508b0deba0aa8db37d97700c26db0396f/test/sphericalModes.jl#L46) and the [TM scattering case](https://github.com/HoBeZwe/SphericalScattering.jl/blob/47393e9508b0deba0aa8db37d97700c26db0396f/test/sphericalModes.jl#L146)).

Concerning, tests like [this](https://github.com/HoBeZwe/SphericalScattering.jl/blob/47393e9508b0deba0aa8db37d97700c26db0396f/test/sphericalModes.jl#L23) we added comments to clearly state the origin of the numerical values and the limited usefulness of the tests.
