### 2023-10-08 #5820 @berquist
My checklist is done.  The revisions to the paper for clarity and the test cleanup are good.  "Values are checked against values determined by this package" shows an understanding of the different kinds of reference tests which is often missing.  I am now convinced enough that the package correctly implements functionality that the paper makes claims about.

The table of [detailed implementation status](https://hobezwe.github.io/SphericalScattering.jl/dev/#Detailed-implementation-status:) that is part of the documentation and not just a GitHub issue is something most packages should have.  The rest of the documentation, both prose and examples, is clear enough to a scientist but non-practitioner.  For the paper itself, there may be some grammar or wording oddities, but I would not nitpick those unless requested since the intent is already clear to me.

I think this paper absolutely belongs in JOSS.
