### 2023-10-12 #5820 @iliarasskazov
I've done with the checklist. I really appreciate the hard work of the authors. Apart from the very implementation of theoretical solution, the proper documentation and organization of the code are even more challenging. The authors done a great job in all of the aspects.

The paper is very much suitable for publication in JOSS.
