### 2023-10-17 #5820 @pdebuyl
@HoBeZwe I will proceed to the [editor's after acceptance tasks](https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance) in the coming days and let you know when an action is required on your side.
