### 2023-10-28 2a41156
replace exp(im*x) by more efficient cis(x)
