### 2023-10-29 #5820 @pdebuyl
Hello @HoBeZwe have started looking at your article + code. Overall it seems to be going ok, except for the comment below on trying the run the code.

I always try to run the basic steps, even as a reviewer, and had trouble in several instance to get the code working. In the basic example, for instance, the plane wave can't serve as the basis for a far-field. Also, I have the error `UndefVarError: `sphericalGridPoints` not defined`.

I installed via conda-forge.

(I'll be "out of office" for a week now, sorry about the delay).

PS: there is a "puplications" typo in the article.
