### 2023-10-29 #5820 @HoBeZwe
Hi, probably the problem here is that you are not using the newest version of the package (hence, the during the review added function sphericalGridPoints does not exist for you). The reason for this is that I have not released a new version so far and the package manager pulls the latest release not the most recent commit.

To run the latest "developer" version you can add the package other than described in the [documentation](https://hobezwe.github.io/SphericalScattering.jl/dev/#Installation) as 'dev SphericalScattering' instead of 'add SphericalScattering'. Then all new features should be available.

Alternatively, I could just release a new version. I just thought that I better do so at the end of the review.

Basically, I would be surprised to see errors occur since, e.g., all of the [Application Examples](https://hobezwe.github.io/SphericalScattering.jl/dev/application/#Application-Examples) are actually executed on GitHub when creating the Documentation. That is, the results in the documentation are obtained by actually running the code.

That the far-field of a plane wave is not defined is intentional as noted in the [documentation](https://hobezwe.github.io/SphericalScattering.jl/dev/planeWave/#Incident-Field).
