### 2023-10-29 bc3711d
fix typo in paper
