+++ .github/workflows/CI.yml
@@ -2,7 +2,7 @@ name: CI
 on:
   push:
     branches:
-      - 'main'
+      - 'master'
       - 'release-'
     tags: '*'
   pull_request:
