+++ versions.js
@@ -1,7 +1,11 @@
 var DOC_VERSIONS = [
   "stable",
+  "v0.5",
+  "v0.4",
+  "v0.3",
+  "v0.2",
   "v0.1",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.1.1";
+var DOCUMENTER_NEWEST = "v0.5.0";
 var DOCUMENTER_STABLE = "stable";
