+++ v0.2.0/joss/paper.bib
@@ -1,305 +1,0 @@
-
-@book{jinTheoryComputationElectromagnetic2015,
-  title     = {Theory and Computation of Electromagnetic Fields},
-  author    = {Jin, Jian-Ming},
-  year      = {2015},
-  edition   = {Second edition},
-  publisher = {{John Wiley \& Sons, Inc}},
-  address   = {{Hoboken, New Jersey}},
-  isbn      = {978-1-119-10804-7}
-}
-
-@book{ruckRadarCrossSection1970,
-  title     = {Radar Cross Section Handbook},
-  author    = {Ruck, George T and Barrick, Donald E and Stuart, William D and Krichbaum, Clarence K},
-  date      = {1970},
-  volume    = {1},
-  publisher = {Plenum Press},
-  address   = {New York},
-  isbn      = {978-1-4899-5326-1}
-}
-
-@article{sihvolaTransmissionLineAnalogy1988,
-  title         = {Transmission Line Analogy for Calculating the Effective Permittivity of Mixtures with Spherical Multilayer Scatterers},
-  author        = {Sihvola, A. and Lindell, I.V.},
-  date          = {1988-01-01},
-  journaltitle  = {Journal of Electromagnetic Waves and Applications},
-  volume        = {2},
-  number        = {8},
-  pages         = {741--756},
-  publisher     = {Taylor \& Francis},
-  url           = {https://www.tandfonline.com/doi/abs/10.1163/156939388X00044},
-}
-
-@book{hansen1988,
-  title     = {Spherical Near-field Antenna Measurements},
-  author    = {Hansen, J. E.},
-  date      = {1988},
-  publisher = {The Institution of Engineering and Technology},
-  address   = {Michael Faraday House, Six Hills Way, Stevenage SG1 2AY, UK},
-  isbn      = {978-0-86341-110-6},
-}
-
-@book{jackson1999,
-  title     = {Classical Electrodynamics},
-  author    = {Jackson, J. D.},
-  date      = {1999},
-  publisher = {Wiley},
-  address   = {New York},
-  number    = {3},
-  isbn      = {0-471-30932-X},
-}
-
-@article{jones1995,
-  title         = {Models for Layered Spherical Particles},
-  author        = {Jones, T. B.},
-  date          = {1995},
-  journaltitle  = {Electromechanics of Particles},
-  pages         = {227–235},
-  publisher     = {Cambridge University Press},
-  doi           = {10.1017/CBO9780511574498}
-}
-
-@article{hofmannExcitationAwareSelfAdaptiveFrequency2023,
-  title         = {An Excitation-Aware and Self-Adaptive Frequency Normalization for Low-Frequency Stabilized Electric Field Integral Equation Formulations},
-  author        = {Hofmann, Bernd and Eibert, Thomas F. and Andriulli, Francesco P. and Adrian, Simon B.},
-  date          = {2023-05},
-  journaltitle  = {IEEE Transactions on Antennas and Propagation},
-  volume        = {71},
-  number        = {5},
-  pages         = {4301--4314},
-  doi           = {10.1109/TAP.2023.3247896}
-}
-
-@article{hofmannLowFrequencyStableExcitation2023,
-  title         = {A Low-Frequency Stable, Excitation Agnostic Discretization of the Right-Hand Side for the Electric Field Integral Equation on Multiply-Connected Geometries},
-  author        = {Hofmann, Bernd and Eibert, Thomas F. and Andriulli, Francesco P. and Adrian, Simon B.},
-  date          = {2023},
-  journaltitle  = {IEEE Transactions on Antennas and Propagation},
[…]
