+++ v0.2.0/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"apiref/#API-Reference","page":"API Reference","title":"API Reference","text":"","category":"section"},{"location":"apiref/","page":"API Reference","title":"API Reference","text":"Modules = [SphericalScattering]","category":"page"},{"location":"apiref/#SphericalScattering.ε0","page":"API Reference","title":"SphericalScattering.ε0","text":"ε0 = 8.8541878176e-12\n\nFree space permittivity.\n\n\n\n\n\n","category":"constant"},{"location":"apiref/#SphericalScattering.μ0","page":"API Reference","title":"SphericalScattering.μ0","text":"μ0 = 4pi * 1e-7\n\nFree space permeability.\n\n\n\n\n\n","category":"constant"},{"location":"apiref/#SphericalScattering.DielectricSphere-Tuple{}","page":"API Reference","title":"SphericalScattering.DielectricSphere","text":"DielectricSphere(\n    radius      = error(\"missing argument `radius`\"), \n    embedding   = Medium(ε0, μ0), \n    filling     = error(\"missing argument `filling`\")\n)\n\nConstructor for the dielectric sphere.\n\n\n\n\n\n","category":"method"},{"location":"apiref/#SphericalScattering.DielectricSphereThinImpedanceLayer-Tuple{}","page":"API Reference","title":"SphericalScattering.DielectricSphereThinImpedanceLayer","text":"DielectricSphereThinImpedanceLayer(\n    radius      = error(\"missing argument `radius`\"),\n    thickness   = error(\"missing argument `thickness` of the coating\"),\n    embedding   = Medium(ε0, μ0),\n    thinlayer   = error(\"missing argument `thinlayer`\"),\n    filling     = error(\"missing argument `filling`\")\n)\n\nConstructor for the dielectric sphere with a thin impedance layer. For this model, it is assumed that the displacement field is only radial direction in the layer, which requires a small thickness and low conductivity. For details, see for example T. B. Jones, Ed., “Models for layered spherical particles,” in Electromechanics of Particles, Cambridge: Cambridge University Press, 1995,  pp. 227–235. doi: 10.1017/CBO9780511574498.012.\n\n\n\n\n\n","category":"method"},{"location":"apiref/#SphericalScattering.FitzgeraldDipole-Tuple{}","page":"API Reference","title":"SphericalScattering.FitzgeraldDipole","text":"ex = FitzgeraldDipole(;\n        embedding   = Medium(ε0, μ0),\n        frequency   = error(\"missing argument `frequency`\"),\n        amplitude   = 1.0,\n        position    = error(\"missing argument `position`\"),\n        orientation = SVector{3,typeof(frequency)}(0.0, 0.0, 1.0),\n)\n\n\n\n\n\n","category":"method"},{"location":"apiref/#SphericalScattering.HertzianDipole-Tuple{}","page":"API Reference","title":"SphericalScattering.HertzianDipole","text":"ex = HertzianDipole(\n        embedding   = Medium(ε0, μ0),\n        frequency   = error(\"missing argument `frequency`\"),\n        amplitude   = 1.0,\n        position    = error(\"missing argument `position`\"),\n        orientation = SVector{3,typeof(frequency)}(0.0, 0.0, 1.0),\n)\n\n\n\n\n\n","category":"method"},{"location":"apiref/#SphericalScattering.LayeredSphere-Tuple{}","page":"API Reference","title":"SphericalScattering.LayeredSphere","text":"LayeredSphere( \n    radii       = error(\"Missing argument `radii`\"), \n    embedding   = Medium(ε0, μ0), \n    filling     = error(\"`missing argument `filling`\")\n)\n\nConstr
[…]
