+++ README.md
@@ -5,7 +5,8 @@
   <img alt="" src="" height="190">
 </picture>
 
-[![Stable](https://img.shields.io/badge/docs-stable-blue.svg)](https://hobezwe.github.io/SphericalScattering.jl/dev/)
+[![Docs-stable](https://img.shields.io/badge/docs-stable-blue.svg)](https://hobezwe.github.io/SphericalScattering.jl/stable/)
+[![Docs-dev](https://img.shields.io/badge/docs-dev-blue.svg)](https://hobezwe.github.io/SphericalScattering.jl/dev/)
 [![MIT license](https://img.shields.io/badge/License-MIT-blue.svg)](https://github.com/HoBeZwe/SphericalScattering.jl/blob/master/LICENSE)
 [![CI](https://github.com/HoBeZwe/SphericalScattering.jl/actions/workflows/CI.yml/badge.svg?branch=master)](https://github.com/HoBeZwe/SphericalScattering.jl/actions/workflows/CI.yml)
 [![codecov](https://codecov.io/gh/HoBeZwe/SphericalScattering.jl/branch/master/graph/badge.svg?token=4F9NUNRC1K)](https://codecov.io/gh/HoBeZwe/SphericalScattering.jl)
@@ -53,4 +54,5 @@ Please cite this package following the information on [Zenodo](https://zenodo.or
 
 ## Documentation
 
-Here you can find the [documentation](https://hobezwe.github.io/SphericalScattering.jl/dev/).
+- Documentation for the [latest stable version](https://hobezwe.github.io/SphericalScattering.jl/stable/).
+- Documentation for the [development version](https://hobezwe.github.io/SphericalScattering.jl/stable/).
