### 2023-11-27 #5820 @pdebuyl
Hello @HoBeZwe

It took me some time to get your code working. With the "dev install", most worked well, thanks for the help.

PlotlyJS on the other hand refused to work. After some googling, it appeared that this library does not support running from a conda install. Even though conda is not an "official end-to-all solutions" it is a convenient one for testing software and the fact that this does not work should be documented.

For information, as you have github repositories in your references, I made sure that they were all save on @softwareheritage to avoid the possible loss of dependencies.

So, my only request is that the incompatibility of the plotting library with conda be stated in the documentation.

Once this is done, please make an archive on zenodo and let me know here of the DOI and version name that you used there. The zenodo entry's authorship and title must match the article.
