+++ Project.toml
@@ -1,7 +1,7 @@
 name = "SphericalScattering"
 uuid = "1a9ea918-b599-4f1f-bd9a-d681e8bb5b3e"
 authors = ["Bernd Hofmann <Bernd.Hofmann@tum.de> and contributors"]
-version = "0.5.0"
+version = "0.6.0"
 
 [deps]
 LegendrePolynomials = "3db4a2ba-fc88-11e8-3e01-49c72059a882"
@@ -18,6 +18,7 @@ SphericalScatteringExt = "PlotlyJS"
 
 [compat]
 LegendrePolynomials = "0.3, 0.4"
+LinearAlgebra = "1"
 PlotlyJS = "0.18"
 SpecialFunctions = "1, 2"
 StaticArrays = "1"
