### 2023-11-28 #5820 @HoBeZwe
Sorry to hear about the inconvenience. Indeed, after a quick test via conda I also got an error when trying to use PlotlyJS.

I added a corresponding warning to the documentation: https://hobezwe.github.io/SphericalScattering.jl/dev/manual/#Plotting-Fields

- The new version is: v0.6.0
- The zenodo DOI: 10.5281/zenodo.10211642
