### 2023-11-28 ebdbbb4
fix typos in JOSS paper
