+++ docs/src/joss/paper.bib
@@ -88,7 +88,7 @@
 }
 
 @inproceedings{hofmannLowFrequencyStableDiscretization2021,
-  title       = {Low-Frequency Stable Discretization of the Electric Field Integral Equation Based on Poincaré's Lemma},
+  title       = {Low-Frequency Stable Discretization of the Electric Field Integral Equation Based on {Poincaré's} Lemma},
   booktitle   = {Proc. IEEE Antennas Propag. Soc. Int. Symp. URSI Nat. Radio Sci. Meeting},
   author      = {Hofmann, Bernd and Eibert, Thomas F. and Andriulli, Francesco P. and Adrian, Simon B.},
   date        = {2021-12},
@@ -107,7 +107,7 @@
 }
 
 @inproceedings{hofmannEfficientCombinationScalarPotential2022a,
-  title       = {Efficient Combination of Scalar-Potential Representations of Solenoidal Functions and Quasi-Helmholtz Projectors},
+  title       = {Efficient Combination of Scalar-Potential Representations of Solenoidal Functions and Quasi-{Helmholtz} Projectors},
   booktitle   = {16th European Conference on Antennas and Propagation (EuCAP)},
   author      = {Hofmann, Bernd and Eibert, Thomas F. and Andriulli, Francesco P. and Adrian, Simon B.},
   date        = {2022-03},
@@ -164,14 +164,14 @@
   number      = {25}, 
   pages       = {691}, 
   author      = {Boaz Blankrot and Clemens Heitzinger}, 
-  title       = {ParticleScattering: Solving and optimizing multiple-scattering problems in Julia}, 
+  title       = {ParticleScattering: Solving and optimizing multiple-scattering problems in {Julia}}, 
   journal     = {Journal of Open Source Software} 
 }
 
 @software{Prahl_miepython_Pure_python,
   author    = {Prahl, Scott},
   doi       = {10.5281/zenodo.7949263},
-  title     = {{miepython: Pure python implementation of Mie scattering}},
+  title     = {{miepython: Pure {Python} implementation of {Mie} scattering}},
   year      = 2023,
   publisher = {Zenodo}
 }
@@ -284,7 +284,7 @@
   author  = {Amos Egel and Lorenzo Pattelli and Giacomo Mazzamuto and Diederik S. Wiersma and Uli Lemmer},
   title   = {{CELES}: {CUDA}-accelerated simulation of electromagnetic scattering by large ensembles of spheres},
   journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
-  year    = {2017},
+  year    = {2017-09},
   volume  = {199},
   pages   = {103--110},
   issn    = {0022-4073},
