### 2023-11-28 #5820 @HoBeZwe
Hi, while reading the proof, we indeed spotted and fixed some typos in the references: Helmholtz, Poincare, and Julia start now with an upper-case letter and we added a comma. With that, everything appears fine from our side :-)
